Move Message methods to extensions.
