+++ A/Discord/Entities/Messages/Extensions/Entity/InteractionMessageExtensions.cs
