+++ A/Discord/Entities/Messages/Extensions/Entity/MessageExtensions.cs
