+++ A/Requests system/Requests/Message/SendInteractionFollowUpMessage/SendInteractionFollowUpHandler.cs
@@ -4,6 +4,7 @@
 
 namespace DiscordLibrary.Requests;
 
+using DiscordLibrary.DiscordEntities;
 using DSharpPlus.Entities;
 using DSharpPlus.Exceptions;
