+++ A/Requests system/Requests/Message/UpdateMessage/UpdateMessageHandler.cs
@@ -4,6 +4,7 @@
 
 namespace RSBingoBot.Requests;
 
+using DiscordLibrary.DiscordEntities;
 using DiscordLibrary.Requests;
 using DSharpPlus.Entities;
